bar chart timeframes working
